chore: move test/utils.ts to testutils.ts
